#451 optimize DeclarativeEnvironmentRecord (#458)
* use MruPropertyCache2 and arguments as stored field
* add missing implicit JsValue operators for uint and char
* simplify often called JsValue Is* As*

+++ Jint/Native/String/StringPrototype.cs
@@ -733,7 +733,7 @@ namespace Jint.Native.String
             {
                 return double.NaN;
             }
-            return s[position];
+            return (double) s[position];
         }
 
         private JsValue CharAt(JsValue thisObj, JsValue[] arguments)
